pc hs - app funcionando

+++ APIs_JAVA/compras/src/main/java/com/compras/api/api/models/Nfe.java
@@ -37,7 +37,7 @@ public class Nfe {
     private float unit;
     private String un;
     private String descricao;
-    private int qtde;
+    private float qtde;
     private String empresa;
     private float vlTotal;
